# openapi/poveventcam.yaml
openapi: 3.0.3
info:
  title: POV EventCamera API
  version: 0.1.0-draft
  description: |
    OpenAPI contract for POV EventCamera guest and organizer APIs.
    Organizer routes support either an HttpOnly organizer session cookie
    or direct Supabase bearer auth (for non-browser clients).
    Guest routes use an HttpOnly guest device session cookie.
servers:
  - url: https://eventpovcamera.app
    description: Organizer web entry domain
  - url: https://guest.eventpovcamera.app
    description: Guest web entry domain
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/organizer/auth/session:
    post:
      tags: [OrganizerAuth]
      summary: Create organizer API session cookie from Supabase bearer token
      operationId: organizerCreateSession
      security:
        - OrganizerBearerAuth: []
      responses:
        '200':
          description: Organizer session created
          headers:
            Set-Cookie:
              description: HttpOnly organizer session cookie scoped to /api/organizer
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerSessionCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [OrganizerAuth]
      summary: Get current organizer session/identity
      operationId: organizerGetSession
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      responses:
        '200':
          description: Organizer session/identity resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerSessionResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [OrganizerAuth]
      summary: Revoke current organizer cookie session (idempotent)
      operationId: organizerDeleteSession
      responses:
        '200':
          description: Session revoked (idempotent)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrganizerSessionDeleteResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events:
    post:
      tags: [OrganizerEvents]
      summary: Create organizer event
      operationId: organizerCreateEvent
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEventRequest'
      responses:
        '200':
          description: Event created or payment redirect required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/CreateEventSuccessResponse'
                  - $ref: '#/components/schemas/PaymentRedirectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: [OrganizerEvents]
      summary: List events for organizer
      operationId: organizerListEvents
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      responses:
        '200':
          description: Events
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/EventSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}:
    get:
      tags: [OrganizerEvents]
      summary: Get event detail and summary stats
      operationId: organizerGetEvent
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event detail
          content:
            application/json:
              schema:
                type: object
                required: [event]
                properties:
                  event:
                    $ref: '#/components/schemas/EventDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [OrganizerEvents]
      summary: Update event settings
      operationId: organizerPatchEvent
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchEventRequest'
      responses:
        '200':
          description: Event updated or payment redirect required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/EventMutationSuccessResponse'
                  - $ref: '#/components/schemas/PaymentRedirectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/close:
    post:
      tags: [OrganizerEvents]
      summary: Close event (stop accepting guest uploads)
      operationId: organizerCloseEvent
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event closed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMutationSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/archive:
    post:
      tags: [OrganizerEvents]
      summary: Archive event
      operationId: organizerArchiveEvent
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Event archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventMutationSuccessResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/gallery:
    get:
      tags: [OrganizerMedia]
      summary: Get paginated media gallery for event
      description: "Filter semantics: within multi-value filters (`filter_uploader`, `filter_tag`) matching is OR; across filter groups matching is AND."
      operationId: organizerGetGallery
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/SortBy'
        - $ref: '#/components/parameters/SortOrder'
        - $ref: '#/components/parameters/FilterDate'
        - $ref: '#/components/parameters/FilterSession'
        - $ref: '#/components/parameters/FilterUploader'
        - $ref: '#/components/parameters/FilterTag'
        - $ref: '#/components/parameters/FilterFileType'
      responses:
        '200':
          description: Gallery page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/gallery/facets:
    get:
      tags: [OrganizerMedia]
      summary: Get searchable uploader/tag facet options for event gallery filters
      operationId: organizerGetGalleryFacets
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/FacetLimit'
        - $ref: '#/components/parameters/UploaderQuery'
        - $ref: '#/components/parameters/TagQuery'
      responses:
        '200':
          description: Facet options
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryFacetsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/gallery/stats:
    get:
      tags: [OrganizerMedia]
      summary: Get event gallery stats
      operationId: organizerGetGalleryStats
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Gallery stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GalleryStatsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/media/{mid}/hide:
    post:
      tags: [OrganizerMedia]
      summary: Hide one media item
      operationId: organizerHideMedia
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/MediaId'
      responses:
        '200':
          description: Media hidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaMutationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/media/{mid}/unhide:
    post:
      tags: [OrganizerMedia]
      summary: Unhide one media item
      operationId: organizerUnhideMedia
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/MediaId'
      responses:
        '200':
          description: Media unhidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaMutationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/media/bulk-hide:
    post:
      tags: [OrganizerMedia]
      summary: Hide multiple media items
      operationId: organizerBulkHideMedia
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkHideRequest'
      responses:
        '200':
          description: Bulk hide completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkHideResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/media/{mid}/download-url:
    get:
      tags: [OrganizerMedia]
      summary: Get signed download URL for original media
      operationId: organizerGetMediaDownloadUrl
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/MediaId'
      responses:
        '200':
          description: Download URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/media/download-urls:
    post:
      tags: [OrganizerMedia]
      summary: Get signed download URLs for multiple media items (max 100)
      operationId: organizerGetMediaDownloadUrls
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadUrlsBatchRequest'
      responses:
        '200':
          description: Batch download URLs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadUrlsBatchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/download-all:
    post:
      tags: [OrganizerMedia]
      summary: Trigger async ZIP job for event downloads
      operationId: organizerDownloadAll
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadAllRequest'
      responses:
        '200':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadAllResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/download-selected:
    post:
      tags: [OrganizerMedia]
      summary: Trigger async ZIP job for selected media downloads
      operationId: organizerDownloadSelected
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadSelectedRequest'
      responses:
        '200':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadAllResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/jobs/{job_id}:
    get:
      tags: [OrganizerJobs]
      summary: Poll async job status
      operationId: organizerGetJobStatus
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/guests:
    get:
      tags: [OrganizerGuests]
      summary: List event guest sessions
      operationId: organizerGetGuests
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
      responses:
        '200':
          description: Guest session list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GuestsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/guests/{sid}/deactivate:
    post:
      tags: [OrganizerGuests]
      summary: Deactivate guest session
      operationId: organizerDeactivateGuest
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeactivateGuestResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/organizer/events/{id}/capacity:
    post:
      tags: [OrganizerEvents]
      summary: Update event capacity and pricing-sensitive settings
      operationId: organizerUpdateCapacity
      security:
        - OrganizerSessionCookie: []
        - OrganizerBearerAuth: []
      parameters:
        - $ref: '#/components/parameters/EventId'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CapacityUpdateRequest'
      responses:
        '200':
          description: Capacity updated or payment redirect required
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/EventMutationSuccessResponse'
                  - $ref: '#/components/schemas/PaymentRedirectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    GuestSessionCookie:
      type: apiKey
      in: cookie
      name: device_session_token
    OrganizerSessionCookie:
      type: apiKey
      in: cookie
      name: organizer_session_token
    OrganizerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    InternalCronBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    EventId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    TooManyRequests:
      description: Rate limited
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    CreateEventSuccessResponse:
      type: object
      required: [event]
      properties:
        event:
          $ref: '#/components/schemas/EventDetail'
...
    OrganizerSessionCreateResponse:
      type: object
      required: [organizer, session]
      properties:
        organizer:
          $ref: '#/components/schemas/OrganizerSessionOrganizer'
        session:
          type: object
          required: [expires_at]
          properties:
            expires_at:
              type: string
              format: date-time
    OrganizerSessionResponse:
      type: object
      required: [organizer, session]
      properties:
        organizer:
          $ref: '#/components/schemas/OrganizerSessionOrganizer'
        session:
          $ref: '#/components/schemas/OrganizerSessionInfo'
    OrganizerSessionDeleteResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          enum: [true]
...